Rename detector to labeler
