+++ packages/firebase_ml_vision/lib/firebase_ml_vision.dart
@@ -15,5 +15,5 @@ import 'package:flutter/services.dart';
 part 'src/barcode_detector.dart';
 part 'src/face_detector.dart';
 part 'src/firebase_vision.dart';
-part 'src/label_detector.dart';
+part 'src/image_labeler.dart';
 part 'src/text_recognizer.dart';
