+++ packages/firebase_ml_vision/lib/src/image_labeler.dart
